requirements.txt
motor==3.6.0
pymongo==4.9.2
python-telegram-bot==20.6
python-telegram-bot[job-queue]
telethon